export id instead of making Node hashable
Making Node hashable is tricky. As the node equality depends on
not only node id, but also node's tree. After parsing, the id can
stay the same but the tree can change, so even if the node is hashable
based on its id, fetching a value from a directory of type [Node:Any]
can lead to surprising missing values.

So the best way so far to id a Node is to export the id field, and let
the user decide how they want to id a node.